moved Maincomand java class to the correct place
